+++ tests/std.scm
@@ -34,3 +34,39 @@
 
         (t.is (vector-map + '#(1 2 3) '#(4 5 6 7))
               '#(5 7 9))))
+
+(test "std: some"
+      (lambda (t)
+        (t.is (some + nil) #f)
+        (t.is (some odd? (list 1 2 3)) #t)
+        (t.is (some odd? (list 2 4 6)) #f)))
+
+(test "std: fold"
+      (lambda (t)
+        (t.is (fold * 1 (cdr (range 10))) 362880)))
+
+(test "std: pluck"
+      (lambda (t)
+        (let ((name (pluck 'name)))
+          (t.is (name 'foo) "foo"))
+        (let ((name (pluck "name")))
+          (t.is (name 'foo) "foo"))
+        (let ((none (pluck)))
+          (t.is (none 'foo) nil))
+        (let ((xy (pluck 'x 'y)))
+          (t.is (xy &(:x 10 :y 20 :z 30)) &(:x 10 :y 20)))))
+
+(test "std: predicates"
+      (lambda (t)
+        (t.is (regex? /foo/) #t)
+        (t.is (boolean? nil) #f)
+        (t.is (boolean? null) #f)
+        (t.is (boolean? undefined) #f)
+        (t.is (boolean? #t) #t)
+        (t.is (boolean? #f) #t)))
+
+(test "std: find"
+      (lambda (t)
+        (t.is (find odd? (list 1 2 3)) 1)
+        (t.is (find odd? (list 0 2 4 3)) 3)
+        (t.is (find odd? (list 0 2 4 6)) nil)))